refactor node and element to implement a ElementsAwareInterface to easily use it for typehinting etc
