+++ src/FeedIo/Feed/Node/ElementInterface.php
@@ -73,42 +73,4 @@ interface ElementInterface
      * @return $this
      */
     public function setAttribute($name, $value);
-
-    /**
-     * returns the ElementIterator to iterate over ElementInterface instances called $name
-     *
-     * @param  string $name element name
-     * @return \FeedIo\Feed\Node\ElementIterator
-     */
-    public function getElementIterator($name);
-
-    /**
-     * returns true if an ElementInterface instance called $name exists
-     *
-     * @param  string $name element name
-     * @return boolean true if the element exists
-     */
-    public function hasElement($name);
-
-    /**
-     * adds $element to the object's attributes
-     *
-     * @param  ElementInterface $element
-     * @return $this
-     */
-    public function addElement(ElementInterface $element);
-
-    /**
-     * Returns all the item's elements
-     *
-     * @return \ArrayIterator
-     */
-    public function getAllElements();
-
-    /**
-     * Returns the item's elements tag names
-     *
-     * @return array
-     */
-    public function listElements();
 }
